requirements.txt
pathway[all]
streamlit
pandas
plotly
watchdog
faker
python-dotenv
openai
